refactor: age default값 설정
